feat: 🚀 actualizar el tipo de dato del campo 'amount' en la entidad de cosecha a 'float8' y en el DTO a 'IsNumber' para permitir valores decimales

+++ src/harvest/entities/harvest-processed.entity.ts
@@ -30,7 +30,7 @@ export class HarvestProcessed {
   })
   harvest: Harvest;
 
-  @Column({ type: 'int4' })
+  @Column({ type: 'float8' })
   amount: number;
 
   @CreateDateColumn()
